In the beginning, Jeffrey Davis created the project.
And the project was without form, and void; and unwritten features
were upon the face of the deep. And Davis said, "Let there be some default information
about this project". And Davis called this information a README.
And Davis said, "give this project a default .gitignore".

And the README and the gitignore was the first commit.